[react-native-webrtc] Adds Config Mod For Android minSdkVersion 24
Summary:
react-native-webrtc requires that the Android minSdkVersion is
set to 24.
Related issue on this here and also mentioned on Daily's RN SDK
that has react-native-webrtc as a dependency.

Adds withBuildScriptExtMinimumVersion same as the ffmpeg-kit-react-native
config plugin. Maybe withBuildScriptExtMinimumVersion could be
exported in `config-plugins` package so that it could be used directly.